+++ tests/test_config.py
@@ -43,26 +43,19 @@ def test_put(bootstrap):
     config.cparser.setValue('textoutput/file', 'invalid2')
     config.cparser.setValue('textoutput/txttemplate', 'invalid3')
 
-    config.put(initialized=True,
-               file='real1',
-               txttemplate='real2',
-               notif=True,
-               loglevel='DEBUG')
+    config.put(initialized=True, notif=True, loglevel='DEBUG')
     del config
     config = bootstrap
 
     assert config.cparser.value('settings/initialized')
     assert config.cparser.value('settings/loglevel') == 'DEBUG'
     assert config.cparser.value('settings/notif')
-    assert config.cparser.value('textoutput/file') == 'real1'
-    assert config.cparser.value('textoutput/txttemplate') == 'real2'
 
 
 def test_get1(bootstrap):
     ''' test basic config.get '''
     config = bootstrap
 
-    assert not config.file
     assert not config.initialized
     assert config.loglevel == 'DEBUG'
     assert not config.notif
@@ -76,12 +69,10 @@ def test_get1(bootstrap):
     config.cparser.setValue('settings/initialized', True)
     config.cparser.setValue('settings/loglevel', 'invalid1')
     config.cparser.setValue('settings/notif', True)
-    config.cparser.setValue('textoutput/file', 'invalid2')
     config.cparser.setValue('textoutput/txttemplate', 'invalid3')
 
     config.get()
 
-    assert config.file == 'invalid2'
     assert config.initialized
     assert config.loglevel == 'invalid1'
     assert config.notif