+++ tests/test_templatehandler.py
@@ -7,6 +7,7 @@ import tempfile
 import pytest
 
 import nowplaying.utils  # pylint: disable=import-error
+import nowplaying.textoutput  # pylint: disable=import-error
 
 
 @pytest.fixture
@@ -33,9 +34,9 @@ def test_writingmeta(gettemplatehandler):  # pylint: disable=redefined-outer-nam
             'title': 'this is the title',
         }
 
-        nowplaying.utils.writetxttrack(filename=filename,
-                                       templatehandler=gettemplatehandler,
-                                       metadata=metadata)
+        nowplaying.textoutput.writetxttrack(filename=filename,
+                                            templatehandler=gettemplatehandler,
+                                            metadata=metadata)
         with open(filename) as tempfn:  # pylint: disable=unspecified-encoding
             content = tempfn.readlines()
 
@@ -51,9 +52,9 @@ def test_missingmeta(gettemplatehandler):  # pylint: disable=redefined-outer-nam
 
         metadata = {}
 
-        nowplaying.utils.writetxttrack(filename=filename,
-                                       templatehandler=gettemplatehandler,
-                                       metadata=metadata)
+        nowplaying.textoutput.writetxttrack(filename=filename,
+                                            templatehandler=gettemplatehandler,
+                                            metadata=metadata)
         with open(filename) as tempfn:  # pylint: disable=unspecified-encoding
             content = tempfn.readlines()
 
@@ -71,9 +72,9 @@ def test_missingtemplate(gettemplatehandler):  # pylint: disable=redefined-outer
             'title': 'this is the title',
         }
 
-        nowplaying.utils.writetxttrack(filename=filename,
-                                       templatehandler=gettemplatehandler,
-                                       metadata=metadata)
+        nowplaying.textoutput.writetxttrack(filename=filename,
+                                            templatehandler=gettemplatehandler,
+                                            metadata=metadata)
         with open(filename) as tempfn:  # pylint: disable=unspecified-encoding
             content = tempfn.readlines()
 
@@ -90,9 +91,9 @@ def test_missingfilename(gettemplatehandler):  # pylint: disable=redefined-outer
             'title': 'this is the title',
         }
 
-        nowplaying.utils.writetxttrack(filename=filename,
-                                       templatehandler=gettemplatehandler,
-                                       metadata=metadata)
+        nowplaying.textoutput.writetxttrack(filename=filename,
+                                            templatehandler=gettemplatehandler,
+                                            metadata=metadata)
         with open(filename) as tempfn:  # pylint: disable=unspecified-encoding
             content = tempfn.readlines()
 
@@ -103,7 +104,7 @@ def test_cleartemplate():  # pylint: disable=redefined-outer-name
     ''' try writing a text '''
     with tempfile.TemporaryDirectory() as newpath:
         filename = os.path.join(newpath, 'test.txt')
-        nowplaying.utils.writetxttrack(filename=filename, clear=True)
+        nowplaying.textoutput.writetxttrack(filename=filename, clear=True)
         with open(filename) as tempfn:  # pylint: disable=unspecified-encoding
             content = tempfn.readlines()
 
@@ -114,7 +115,7 @@ def test_justafile():  # pylint: disable=redefined-outer-name
     ''' try writing a text '''
     with tempfile.TemporaryDirectory() as newpath:
         filename = os.path.join(newpath, 'test.txt')
-        nowplaying.utils.writetxttrack(filename=filename)
+        nowplaying.textoutput.writetxttrack(filename=filename)
         with open(filename) as tempfn:  # pylint: disable=unspecified-encoding
             content = tempfn.readlines()